guiファイルに切り出し
