+++ test.html
@@ -79,10 +79,10 @@
                     ダメージの最大値と最小値も表示する(HPの計算をしない)
                 </label>
             </div>
-            <div id="output"></div>
             <div class="right">
                 <button id="run_turn" disabled>ターンを進める</button>
             </div>
+            <div id="output"></div>
         </div>
     </div>
 </body>
